# Server/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "BlueWave Uptime",
		"summary": "BlueWave Uptime OpenAPI Specifications",
		"description": "BlueWave Uptime is an open source server monitoring application used to track the operational status and performance of servers and websites. It regularly checks whether a server/website is accessible and performs optimally, providing real-time alerts and reports on the monitored services' availability, downtime, and response time.",
		"contact": {
			"name": "API Support",
			"url": "mailto:support@bluewavelabs.ca",
			"email": "support@bluewavelabs.ca"
		},
		"license": {
			"name": "AGPLv3",
			"url": "https://github.com/bluewave-labs/bluewave-uptime/tree/HEAD/LICENSE"
		},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://localhost:{PORT}/{API_PATH}",
			"description": "Local Development Server",
			"variables": {
				"PORT": {
					"description": "API Port",
					"enum": [
						"5000"
					],
					"default": "5000"
				},
				"API_PATH": {
					"description": "API Base Path",
					"enum": [
						"api/v1"
					],
					"default": "api/v1"
				}
			}
		},
		{
			"url": "http://localhost/{API_PATH}",
			"description": "Distribution Local Development Server",
			"variables": {
				"API_PATH": {
					"description": "API Base Path",
					"enum": [
						"api/v1"
					],
					"default": "api/v1"
				}
			}
		},
		{
			"url": "https://uptime-demo.bluewavelabs.ca/{API_PATH}",
			"description": "Bluewave Demo Server",
			"variables": {
...
	"paths": {
...
		"/invite": {
			"post": {
				"tags": [
					"invite"
				],
				"description": "Request an invitation",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"email",
									"role"
								],
								"properties": {
									"email": {
										"type": "string"
									},
									"role": {
										"type": "array"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
...
		"/monitors": {
			"get": {
				"tags": [
					"monitors"
				],
				"description": "Get all monitors",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"post": {
				"tags": [
					"monitors"
				],
				"description": "Create a new monitor",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMonitorBody"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"delete": {
				"tags": [
					"monitors"
				],
				"description": "Delete all monitors",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"default": false
					},
					"msg": {
						"type": "string"
					}
				}
			},
			"SuccessResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"default": true
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"type": "object"
					}
				}
			},
...
			"CreateMonitorBody": {
				"type": "object",
				"required": [
					"userId",
					"teamId",
					"name",
					"description",
					"type",
					"url"
				],
				"properties": {
					"_id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"teamId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"http",
							"ping",
							"pagespeed"
						]
					},
					"url": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"interval": {
						"type": "integer"
					},
					"notifications": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
...
		}
	}
}